requirements.txt
affine==2.1.0
amqp==2.2.2
appdirs==1.4.3
asn1crypto==0.24.0
astroid==1.5.3
awsebcli==3.12.1
backports.ssl-match-hostname==3.5.0.1
bcrypt==3.1.4
billiard==3.5.0.3
blessed==1.14.2
boto==2.48.0
boto3==1.4.7
botocore==1.8.22
celery==4.1.0
cement==2.8.2
cffi==1.11.2
click==6.7
click-plugins==1.0.3
cligj==0.4.0
colorama==0.3.7
cryptography==2.1.4
cycler==0.10.0
Django==1.10.6
django-ipware==1.1.6
django-simple-history==1.9.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
Fabric==1.14.0
futures==2.2.0
GDAL==2.2.3
idna==2.6
isort==4.2.15
Jinja2==2.9.6
jmespath==0.9.3
keyring==10.3.2
keyrings.alt==2.2
kombu==4.1.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.6.1
numpy==1.13.3
opencv-python==3.3.0.10
packaging==16.8
paramiko==2.4.0
pathspec==0.5.0
planet==1.1.0
prettytable==0.7.2
psycopg2==2.7.1
pyasn1==0.2.3
pycparser==2.18
PyJWT==1.5.3
pylint==1.7.2
PyNaCl==1.2.1
pyparsing==2.2.0
PySocks==1.6.7
python-dateutil==2.6.1
python-fmask==0.4.5
python-magic==0.4.13
pytz==2017.2
PyYAML==3.12
rasterio==0.36.0
redis==2.10.6
requests==2.9.1
requests-futures==0.9.7
retrying==1.3.3
rios==1.4.4
rsa==3.4.2
s3transfer==0.1.11
scipy==1.0.0
semantic-version==2.5.0
six==1.11.0
snuggs==1.4.1
tabulate==0.7.5
termcolor==1.1.0
twilio==6.3.0
ua-parser==0.7.3
user-agents==1.1.0
vine==1.1.4
virtualenv==15.1.0
wcwidth==0.1.7
websocket-client==0.46.0
wrapt==1.10.11
